requirements.txt
pyserial
sphinx
wheel
twine
pytest